runtime: split parts of fd_runtime.c into fd_runtime_init.c
Moves save/restore related functionality into a separate compile
unit.  This allows using snapshot functionality without depending
on the rest of the runtime.
