+++ src/flamenco/runtime/fd_runtime.h
@@ -3,6 +3,7 @@
 
 #include "../fd_flamenco_base.h"
 #include "fd_runtime_err.h"
+#include "fd_runtime_init.h"
 #include "fd_rocksdb.h"
 #include "fd_acc_mgr.h"
 #include "../features/fd_features.h"
@@ -219,31 +220,9 @@ ulong
 fd_runtime_lamports_per_signature_for_blockhash( fd_exec_slot_ctx_t const * slot_ctx,
                                                  fd_hash_t const * blockhash );
 
-fd_funk_rec_key_t
-fd_runtime_firedancer_bank_key( void );
-
-fd_funk_rec_key_t
-fd_runtime_epoch_bank_key( void );
-
-fd_funk_rec_key_t
-fd_runtime_slot_bank_key( void );
-
-int
-fd_runtime_save_slot_bank( fd_exec_slot_ctx_t * slot_ctx );
-
-int
-fd_runtime_save_epoch_bank( fd_exec_slot_ctx_t * slot_ctx );
-
 // int
 // fd_global_import_solana_manifest( fd_exec_slot_ctx_t * slot_ctx,
 //                                   fd_solana_manifest_t * manifest);
-
-/* fd_features_restore loads all known feature accounts from the
-   accounts database.  This is used when initializing bank from a
-   snapshot. */
-
-void
-fd_features_restore( fd_exec_slot_ctx_t * slot_ctx );
 
 static inline ulong
 fd_rent_exempt( fd_rent_t const * rent,
@@ -261,17 +240,6 @@ fd_runtime_update_leaders( fd_exec_slot_ctx_t * slot_ctx, ulong slot );
 /* rollback runtime to the state where the given slot just FINISHED executing */
 int
 fd_runtime_rollback_to( fd_exec_slot_ctx_t * slot_ctx, ulong slot );
-
-/* Recover slot_bank and epoch_bnck from funky */
-void
-fd_runtime_recover_banks( fd_exec_slot_ctx_t * slot_ctx, int delete_first );
-
-void
-fd_runtime_delete_banks( fd_exec_slot_ctx_t * slot_ctx );
-
-/* Recover slot_ctx from funky */
-void
-fd_runtime_recover_slot_ctx( fd_exec_slot_ctx_t * slot_ctx );
 
 /* fd_runtime_ctx_{align,footprint} return FD_REPLAY_STATE_{ALIGN,FOOTPRINT}. */
